openDeleteDatabaseConfirmationPane
